chg: move key value module to modules folder.
