+++ src/app/modules/key-value/key-value-list/key-value-list.component.scss
