+++ src/app/modules/key-value/key-value.interface.ts
