+++ src/app/modules/key-value/key-value.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { KeyValueComponent } from './key-value/key-value.component';
 import { KeyValueFormComponent } from './key-value-form/key-value-form.component';
 import { KeyValueListComponent } from './key-value-list/key-value-list.component';
-import { SharedModule } from '../../../../shared/shared.module';
+import { SharedModule } from '../../shared/shared.module';
 
 
 @NgModule({
